refactor: 프로필 메뉴 분리
